requirements.txt
# GIL Laboratorios - Dependencias
# Sistema de Gestión Integral de Laboratorios SENA

# Framework Web
Flask==3.1.2
flask-cors==6.0.2
Werkzeug==3.1.4

# Base de datos
mysql-connector-python==9.5.0

# Autenticación
PyJWT==2.10.1
bcrypt==5.0.0

# Variables de entorno
python-dotenv==1.2.1

# Procesamiento de imágenes
Pillow==12.0.0
opencv-python==4.12.0.88
numpy==2.2.6

# Machine Learning / IA
scikit-learn==1.8.0
joblib==1.5.2
tensorflow==2.20.0

# Reconocimiento de voz
pydub==0.25.1

# Código QR
qrcode==8.2

# Utilidades
requests==2.32.5
